requirements.txt
opencv-python==4.7.0.72
numpy==1.23.5
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-yasg==1.21.5
keras==2.10.0
Keras-Preprocessing==1.1.2
pandas==1.5.3
tensorflow==2.10.0